Get decomposing nodes

Get attribute values from the nodes that are decomposing a node (e.g. the growing units decomposing an axis).

Get attribute values from the nodes that are decomposing a node (e.g. the growing units decomposing an axis).

decompose(
  attribute,
  node = NULL,
  scale = NULL,
  symbol = NULL,
  filter_fun = NULL
)

decompose(
  attribute,
  node = NULL,
  scale = NULL,
  symbol = NULL,
  filter_fun = NULL
)

Arguments

attribute
Any node attribute (as a character)

node
The MTG node

scale
An integer vector for filtering the .scale of the decomposing nodes (i.e. the SCALE column from the MTG classes).

symbol
A character vector for filtering the decomposing nodes by the name of their .symbol (i.e. the SYMBOL column from the MTG classes).

filter_fun
Any filtering function taking a node as input.

Value

The attribute values from the nodes decomposing the input node

The attribute values from the nodes decomposing the input node

Details

This function is mainly used to summarize attributes at a higher scale when they were measured at a lower scale. For example we can think of an mtg where the length was measured at the internode scale only, so this function can be used to summarize it at e.g. axis scale. The filters are used to get the nodes of interest only. For example an axis can be made of internodes and leaves, but the length of an axis is made from the cumulative length of the internodes only (not the leaves). A second example would be to summarize the total leaf area per axis, where we would only want to use the area of the leaves (not the internodes).

This function is mainly used to summarize attributes at a higher scale when they were measured at a lower scale. For example we can think of an mtg where the length was measured at the internode scale only, so this function can be used to summarize it at e.g. axis scale. The filters are used to get the nodes of interest only. For example an axis can be made of internodes and leaves, but the length of an axis is made from the cumulative length of the internodes only (not the leaves). A second example would be to summarize the total leaf area per axis, where we would only want to use the area of the leaves (not the internodes).

Note

This function is mainly intended to be used with mutate_mtg(). In this case, the node argument can be left empty (or you can put node = node equivalently).

This function is mainly intended to be used with mutate_mtg(). In this case, the node argument can be left empty (or you can put node = node equivalently).

Examples

filepath= system.file("extdata", "simple_plant_3.mtg", package = "XploRer")
MTG = read_mtg(filepath)

decompose(".symbol", node = extract_node(MTG,"node_3"), decomp_type = "symbol")

#>      node_4      node_5      node_9 
#> "Internode"      "Leaf" "Internode" 

# using filters to remove nodes that we don't need:
decompose(".symbol", node = extract_node(MTG,"node_3"), symbol = "Internode", decomp_type = "symbol")

#>      node_4      node_9 
#> "Internode" "Internode" 

# We can check the function worked well by printing the symbols of each nodes:
print(MTG$MTG, ".symbol")

#>                              levelName    .symbol
#> 1  node_1                                   Scene
#> 2   °--node_2                          Individual
#> 3       °--node_3                            Axis
#> 4           °--node_4                   Internode
#> 5               ¦--node_5                    Leaf
#> 6               ¦--node_6                    Axis
#> 7               ¦   °--node_7           Internode
#> 8               ¦       °--node_8            Leaf
#> 9               °--node_9               Internode
#> 10                  °--node_10               Axis
#> 11                      °--node_11      Internode
#> 12                          °--node_12       Leaf

filepath= system.file("extdata", "simple_plant_3.mtg", package = "XploRer")
MTG = read_mtg(filepath)

axis(attribute = "length", node = extract_node(MTG,"node_3"), symbol = "Leaf")

#> Error in axis(attribute = "length", node = extract_node(MTG, "node_3"),     symbol = "Leaf"): argument "side" is missing, with no default

axis(attribute = "name", node = extract_node(MTG,"node_3"), symbol = "Internode")

#> Error in axis(attribute = "name", node = extract_node(MTG, "node_3"),     symbol = "Internode"): argument "side" is missing, with no default

# The MTG has two leaves (node_5 and node_7):
leaves("Length", node = MTG$MTG)

#>  node_5  node_8 node_12 
#>      NA      10      10 

# We can check the function worked well by printing the symbols of each nodes:
print(MTG$MTG, ".symbol")

#>                              levelName    .symbol
#> 1  node_1                                   Scene
#> 2   °--node_2                          Individual
#> 3       °--node_3                            Axis
#> 4           °--node_4                   Internode
#> 5               ¦--node_5                    Leaf
#> 6               ¦--node_6                    Axis
#> 7               ¦   °--node_7           Internode
#> 8               ¦       °--node_8            Leaf
#> 9               °--node_9               Internode
#> 10                  °--node_10               Axis
#> 11                      °--node_11      Internode
#> 12                          °--node_12       Leaf

# If we need the terminal Internodes:
leaves("Length", node = MTG$MTG, symbol = "Internode")

#>  node_7 node_11 
#>       4       4 

# Or the leaves at a given scale:
print(MTG$MTG, ".scale")

#>                              levelName .scale
#> 1  node_1                                  NA
#> 2   °--node_2                               1
#> 3       °--node_3                           2
#> 4           °--node_4                       3
#> 5               ¦--node_5                   3
#> 6               ¦--node_6                   2
#> 7               ¦   °--node_7               3
#> 8               ¦       °--node_8           3
#> 9               °--node_9                   3
#> 10                  °--node_10              2
#> 11                      °--node_11          3
#> 12                          °--node_12      3

leaves("Length", node = MTG$MTG, scale = 2)

#>  node_6 node_10 
#>      NA      NA