Get decomposing nodes recursively

This is the function actually doing the hard work for [decompose()].

decompose_(
  attribute,
  node = NULL,
  ref_node = NULL,
  scale = NULL,
  symbol = NULL,
  link = NULL,
  filter_fun = NULL,
  decomp_type = c("scale", "symbol")
)

Arguments

attribute
Any node attribute (as a character)

node
The MTG node

ref_node
The reference node to match the scale or symbol

scale
An integer vector for filtering the .scale of the decomposing nodes (i.e. the SCALE column from the MTG classes).

symbol
A character vector for filtering the decomposing nodes by the name of their .symbol (i.e. the SYMBOL column from the MTG classes).

link
A character vector for filtering the .link with the descendant

filter_fun
Any filtering function taking a node as input.

decomp_type
The type of scale to use for decomposition, either the scale or the symbol of the node.

Value

The attributes of the nodes decomposing the reference node.