Get ancestors value

Get attribute values from all ancestors (all nodes above).

get_ancestors_values(attribute, node = NULL, scale = NULL, self = FALSE)

Arguments

attribute
Any node attribute (as a character)

node
The MTG node

scale
Filter ancestors by the names of the MTG scale(s) (i.e. the SYMBOL from the MTG classes).

self
Return the value of the current node (TRUE), or the ancestors only (FALSE, the default)

Value

The attribute values from the ancestors of the node (from first parent to farther ancestor)

Details

This function is mainly intended to be used with mutate_mtg(). In this case, the node argument can be left empty (or you can put node = node equivalently).

Examples

filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG = read_mtg(filepath)

# node_6 has four ancestors:
get_ancestors_values("Length", node = data.tree::FindNode(MTG$MTG, "node_6"))

#> node_5 node_3 node_2 node_1 
#>      6      4     NA     NA 

# Two of them have no values for Length

# If the value of node_6 is also needed:
get_ancestors_values("Length", node = data.tree::FindNode(MTG$MTG, "node_6"), self = TRUE)

#> node_6 node_5 node_3 node_2 node_1 
#>     12      6      4     NA     NA 

# We can filter by scale if we need to return the values for some scales only:
get_ancestors_values("Width", node = data.tree::FindNode(MTG$MTG, "node_6"), scale = "Internode")

#> node_5 node_3 
#>     NA      1 

# The values are only returned for the ancestors with the required scale.
# For example we know that a leaf cannot be an ancestor because it cannot bear anything:
get_ancestors_values("Width", node = data.tree::FindNode(MTG$MTG, "node_6"), scale = "Leaf")

#> logical(0)

# In this case it returns a length 0 vector.

# Here we get the value of node_6 also, because its parent "node_5" is not of scale
# "Leaf", so it was filtered out. It you need the values for one scale, but not
# making a recursive search from one scale to another until finding the required scale,
# you can put the `recursive` argument to `FALSE`: