Get children value

Get attribute values from the children of a node.

get_children_values(attribute, node = NULL, scale = NULL, recursive = TRUE)

Arguments

attribute
Any node attribute (as a character)

node
The MTG node

scale
The names of the MTG scale(s) required for the children (i.e. the SYMBOL from the MTG classes). Used as a filter.

recursive
If a child is not of the right scale, continue until the scale required is met if TRUE, or returns NA if FALSE.

Value

The attribute values from the children of the node

Details

This function is mainly intended to be used with mutate_mtg(). In this case, the node argument can be left empty (or you can put node = node equivalently).

Examples

filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG = read_mtg(filepath)

# node_5 has one child:
get_children_values("Length", node = data.tree::FindNode(MTG$MTG, "node_5"))

#> node_6 
#>     12 

# Using node 3 as reference now:
node_3 = data.tree::FindNode(MTG$MTG, "node_3")
# node_3 has two children, returns two values:
get_children_values("Length", node = node_3)

#> node_4 node_5 
#>     10      6 

# To get the names of those children:
get_children_values("name", node = node_3)

#>   node_4   node_5 
#> "node_4" "node_5" 

# The width is not available for one child ("node_5"):
get_children_values("Width", node = node_3)

#> node_4 node_5 
#>      6     NA 

# We can filter by scale if we need to return the values for some scales only:
get_children_values("Width", node = node_3, scale = "Leaf")

#> node_4 node_6 
#>      6      7 

# Here we get the value of node_6 also, because its parent "node_5" is not of scale
# "Leaf", so it was filtered out. It you need the values for one scale, but not
# making a recursive search from one scale to another until finding the required scale,
# you can put the `recursive` argument to `FALSE`:

# We can also get the values recursively until finding the right value:
get_children_values("Width", node = node_3, scale = "Leaf", recursive = FALSE)

#> node_4        
#>      6     NA 

# To get the values of the children of each node:
mutate_mtg(MTG, children_width = get_children_values("Width"))
print(MTG$MTG, "Width", "children_width")

#>                levelName Width children_width
#> 1 node_1                    NA               
#> 2  °--node_2                NA              1
#> 3      °--node_3             1          6, NA
#> 4          ¦--node_4         6               
#> 5          °--node_5        NA              7
#> 6              °--node_6     7