Apply function to tree

Apply function to tree

mtg_apply(
  data,
  fun,
  scale,
  traversal = c("pre-order", "post-order", "in-order", "level", "ancestor")
)

Arguments

data
The input MTG, as from read_MTG()

fun
The function to apply (see details)

scale
The scale to apply the function to.

traversal
The traversal order of the tree. Can be any of 'pre-order', 'post-order', 'in-order', 'level', 'ancestor', or a custom function (see details)

Value

Nothing. The MTG is modified in place.

Details

The traversal is as in data.tree::Traverse(). Here's its documentation:

The traversal order is as follows. (Note that these descriptions are not precise and complete. They are meant for quick reference only. See the data.tree vignette for a more detailed description).

pre-order: Go to first child, then to its first child, etc.

post-order: Go to the first branch's leaf, then to its siblings, and work your way back to the root

in-order:Go to the first branch's leaf, then to its parent, and only then to the leaf's sibling

level: Collect root, then level 2, then level 3, etc.

ancestor: Take a node, then the node's parent, then that node's parent in turn, etc. This ignores the pruneFun

function: You can also provide a function, whose sole parameter is a Node object. The function is expected to return the node's next node, a list of the node's next nodes, or NULL.

Note

The filter on scale is supposed to happen in fun. This approach gives the user more freedom.

Examples

filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG = read_MTG(filepath)

# Defining the function to apply, here we compute the cross-section surface:

Width

#> Error in eval(expr, envir, enclos): object 'Width' not found

fun = function(node) {
node$section_surface_parent = get_parent_value("Length","Width",  node = node)
}

# Applying the function to the MTG:
  mtg_apply(MTG = MTG, fun = fun)

#> Error in mtg_apply(MTG = MTG, fun = fun): unused argument (MTG = MTG)

# Checking it worked:
  mtg_df = ToDataFrameTree(mtg$MTG,"length","diameter","topological_order","section_surface","section_surface_childs","section_surface_parent")

#> Error in ToDataFrameTree(mtg$MTG, "length", "diameter", "topological_order",     "section_surface", "section_surface_childs", "section_surface_parent"): could not find function "ToDataFrameTree"